[zircon] Rework the riscv exception entry code and percpu struct
This rebuilds our exception entry code with a similar logic to the one
used in littlekernel: user exceptions and kernel exceptions are two
distinct functions expanded from shared macros, this shifts some of the
current runtime check into "compile time checks".

This also redefines the percpu convention to store the percpu pointer in
the gp register instead of t6.

Change-Id: I1ed816eff961c0e05ca49673bd2c063d163dd363
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538963
Reviewed-by: Roger Tawa <rogerta@google.com>