[zircon] Handle context switch to kernel aspace on riscv
Until now, when switching from userspace threads to kernelspace threads,
we would keep the page tables of the userspace threads. This happened to
work in most cases because the userspace threads would also map high
memory, until the userspace thread would die and its page tables would
be recycled by something else corrupting the kernel thread.

We first need to save the physical address of the kernel page tables and
restore it on ContextSwitch calls with new_aspace == NULL.

This was due to an unfortunate copy/paste from the arm64 platform code
where there are separate registers for high memory mapping and low
memory mapping.

Change-Id: I7b7c544ca6107b9932970588e402b9bec6eac78c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553626
Reviewed-by: Roger Tawa <rogerta@google.com>