[zircon] HACK: Replace the BrwLock in the handle table with a Mutex
It's unclear whether the BrwLock implementation has an issue on riscv
(it uses a 128 bit "atomic" store which is implemented by compiler-rt as
64 bit stores guarded by a 64 bit atomic spinlock) or if replacing it
with a Mutex hides a concurency bug but this hack lets us boot on multi
CPUs up to a shell.

We can't reasonably keep this long-term because it butchers performances
but well, it boots. ¯\_(ツ)_/¯

Change-Id: Iaa08f3cfd034f3fe92bc92526ea44db881345d8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553624
Reviewed-by: Roger Tawa <rogerta@google.com>